Show escaped characters only in visual field
